requirements.txt
Flask==3.1.0
blinker==1.9
tensorflow-cpu==2.12.1
keras==2.12.0
numpy==1.24.3
pandas==1.5.3
pillow==9.5.0